refactor FastHealthyMealsUIController for readability and structure
