+++ src/main/kotlin/logic/usecase/GetFastHealthyMealsUseCase.kt
@@ -1,6 +1,7 @@
-package logic
+package logic.usecase
 
 
+import logic.MealsRepository
 import logic.exception.FoodMoodException
 import logic.models.Meal
